Added Base Ramsete
Added all of the code needed to make ramsete work.

Now its just a matter of showing how to construct trajectories.

I wanted this to be stripped down so it could explain the key parts of Ramsete but I'm realizing I need to test this on a robot so I need to add stuff in.