Bus Voltage + Comments
Added a ton of comments explaining different config values.

Made some changes because I googled what talon considers 100% voltage. I'll copy paste this that I wrote in comments.

ArbFF accepts a value from -1 to 1. This is a percentage output to apply. It considers 100% (or a value of 1) to be the voltage coming into the TalonSRX. So to give the motor the voltage calculated from SimpleMotorFeedforward, SimpleMotorFeedforward's value is divided by the bus voltage, then the talon multiplies it by the bus voltage.