Experimenting with an idea
I took the concept that the talons use to run motion magic to drive the robot accurate forward using differential sensor sum and using gyro data to steer the robot.

I "split the motion magic apart" in my version. The velocity pid is run by the robot characterization data and the motion profile is created by the TrapizoidProfile object.